implement two pass parsing
This splits the parsing process into two passes: scanning and parsing.
The scanner converts lines into a tree of nodes of various line types
and then they are parsed into the final output structs. This allows us
to use continuations as expected by the BAI2 spec.

It also modifies the output to output domain objects rather than
exposing all the implementation details of BAI2 files. This should be
more useful for understanding what the file contains.